Fix three defects the warning sweep exposed
Archiver::testArchiveAndGetSize returned uint64_t and signalled failure with
"return -1", which wraps to UINT64_MAX. Every caller tested "archiveSize < 0",
which an unsigned value can never satisfy, so the corrupt-archive check has
never fired: a truncated or damaged download went straight to extraction. The
size is now int64_t end to end -- function, signals, slots -- so the existing
checks mean what they say.

Settings could dereference an uninitialised pointer. The resolution loop walks
every entry in INGAME_RES, a space-separated list the user can extend, but the
dialog has three rows and the switch selecting them had cases 0-2 and no
default. A fourth resolution left both combo-box pointers uninitialised and used
them immediately: opening Settings was undefined behaviour. Both loops now
initialise to nullptr and stop when they run out of rows.

Uninstall reported nothing when it failed. A rename that did not happen was
simply left out of the moved list, so the item stayed installed, no message
appeared, and the list refreshed as if it had worked -- and a partial failure
left half an add-on behind with a manifest describing only the half that moved.
Failures are now named, and nothing that moved nothing writes a manifest.

Also adds the missing START_WITHOUT_MODS case, which reported "Unknown start
type" for a menu entry that works.

+++ src/archiver.cpp
@@ -114,7 +114,7 @@ bool Archiver::compressSingleFile(QFile *inputFile, std::string outputPath)
     return false;
 }
 
-uint64_t Archiver::testArchiveAndGetSize(QFile *archiveFile)
+int64_t Archiver::testArchiveAndGetSize(QFile *archiveFile)
 {
     // Get file info for the archive file
     QFileInfo archiveFileInfo(archiveFile->filesystemFileName());
@@ -131,11 +131,14 @@ uint64_t Archiver::testArchiveAndGetSize(QFile *archiveFile)
         archive.test();
 
         // Return the total size of the uncompressed files
-        return archive.size();
+        return static_cast<int64_t>(archive.size());
 
     } catch (const bit7z::BitException& ex) {
 
         qWarning() << "Archive test failure:" << ex.what();
+        // Signed on purpose: this used to be uint64_t, so -1 wrapped to UINT64_MAX
+        // and every caller's "if (archiveSize < 0)" was always false. A corrupt or
+        // truncated archive sailed straight through to extraction.
         return -1;
     }
 }
